move render test out of runtime
rename
