+++ src/Entry/EGL/CMakeLists.txt
@@ -2,8 +2,3 @@ add_library(egl_runtime EGLRuntime.cpp)
 set_target_properties(egl_runtime PROPERTIES POSITION_INDEPENDENT_CODE ON)
 target_link_libraries(egl_runtime vgg_application)
 target_link_libraries(egl_runtime OpenGL::EGL)
-add_executable(egl_entry EGLEntry.cpp)
-target_link_libraries(egl_entry egl_runtime)
-target_include_directories(egl_entry PRIVATE ${VGG_CONTRIB_ARGPARSE_INCLUDE})
-get_target_property(VGG_LAYER_RPATH vgg_layer BUILD_RPATH)
-set_target_properties(egl_entry PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR} BUILD_RPATH ${VGG_LAYER_RPATH})
