--- src/Layer/interface/Reader/AiFileReader.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "Common/Config.h"
-#include "IReader.hpp"
-#include <filesystem>
-namespace VGG
-{
-
-class VGG_EXPORTS AiFileReader : public IReader
-{
-};
-} // namespace VGG
