--- src/Layer/interface/Reader/FigmaFileReader.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "Common/Config.h"
-#include "IReader.hpp"
-#include <filesystem>
-namespace VGG
-{
-
-class VGG_EXPORTS FigmaFileReader : public IReader
-{
-};
-} // namespace VGG
