--- src/Layer/interface/Reader/SketchFileReader.h
@@ -1,10 +1,0 @@
-#pragma once
-#include "Common/Config.h"
-#include "IReader.hpp"
-namespace VGG
-{
-
-class VGG_EXPORTS SketchFileReader : public IReader
-{
-};
-} // namespace VGG
